created new workspace to remove clean up migrations files and add default values using migration files. removed cart_items_test table from migration and relevant js files

+++ migrations/20220823064830-orders.js
@@ -23,7 +23,7 @@ exports.up = function(db) {
       autoIncrement:true
     },
     shipping_address:{type:'string',length:100,notNull:true},
-    postal_code:{type:'smallint',unsigned:true,notNull:true},
+    postal_code:{type:'int',unsigned:true,notNull:true},
     remarks:{type:'string',length:'200',notNull:true, defaultValue:'nil'},
     datetime_created:{type:'datetime', notNull:true},
     datetime_last_modified:{type:'datetime', notNull:true},
